refactor: Organize data as feature
