+++ lib/mdigits.dart
@@ -1,5 +1,5 @@
 library mdigits;
 
 export 'src/mdigits_activity.dart';
-export 'src/components/models/mdigits_data.dart';
+export 'src/components/data/mdigits_data.dart';
 export 'src/components/config/user_config.dart';
