+++ lib/src/components/activity/activity_view.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
-import 'package:mdigits/src/data.dart';
+import 'package:mdigits/src/components/data/data.dart';
 import 'package:mdigits/src/components/end/end_view.dart';
 import 'package:mdigits/src/components/instructions/instructions_view.dart';
 import 'package:mdigits/src/components/activity/activity_controller.dart';
