+++ lib/src/components/response/response_controller.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
-import 'package:mdigits/src/data.dart';
+import 'package:mdigits/src/components/data/data.dart';
 import 'package:mdigits/src/components/activity/activity_controller.dart';
 import 'package:mdigits/src/components/stim/stim_controller.dart';
 import 'package:mdigits/src/components/config/app_config.dart';
